:recycle: Remove transaction folder and move to root
