+++ packages/lisk-transactions/src/0_transfer.ts
@@ -13,9 +13,9 @@
  *
  */
 import { getAddressFromPublicKey } from '@liskhq/lisk-cryptography';
+import { TransferTransaction } from './0_transfer_transaction';
 import { BYTESIZES, TRANSFER_FEE } from './constants';
 import { TransactionJSON } from './transaction_types';
-import { TransferTransaction } from './transactions';
 import {
 	createBaseTransaction,
 	validateAddress,
