+++ packages/lisk-transactions/src/0_transfer_transaction.ts
@@ -14,16 +14,15 @@
  */
 import { getAddressFromPublicKey } from '@liskhq/lisk-cryptography';
 import * as BigNum from 'browserify-bignum';
-import { MAX_TRANSACTION_AMOUNT, TRANSFER_FEE } from '../constants';
-import { TransactionError, TransactionMultiError } from '../errors';
-import { Account, TransactionJSON, TransferAsset } from '../transaction_types';
 import {
-	validateAddress,
-	validateTransferAmount,
-	validator,
-	verifyBalance,
-} from '../utils';
-import { BaseTransaction, StateStore, StateStorePrepare } from './base';
+	BaseTransaction,
+	StateStore,
+	StateStorePrepare,
+} from './base_transaction';
+import { MAX_TRANSACTION_AMOUNT, TRANSFER_FEE } from './constants';
+import { TransactionError, TransactionMultiError } from './errors';
+import { Account, TransactionJSON, TransferAsset } from './transaction_types';
+import { validateAddress, validateTransferAmount, validator, verifyBalance } from './utils';
 
 const TRANSACTION_TRANSFER_TYPE = 0;
