+++ packages/lisk-transactions/src/1_register_second_passphrase.ts
@@ -13,9 +13,9 @@
  *
  */
 import { getKeys } from '@liskhq/lisk-cryptography';
+import { SecondSignatureTransaction } from './1_second_signature_transaction';
 import { SIGNATURE_FEE } from './constants';
 import { TransactionJSON } from './transaction_types';
-import { SecondSignatureTransaction } from './transactions';
 import { createBaseTransaction } from './utils';
 
 export interface SecondPassphraseInputs {
