+++ packages/lisk-transactions/src/1_second_signature_transaction.ts
@@ -14,11 +14,15 @@
  */
 import { hash, hexToBuffer, signData } from '@liskhq/lisk-cryptography';
 import * as BigNum from 'browserify-bignum';
-import { SIGNATURE_FEE } from '../constants';
-import { TransactionError, TransactionMultiError } from '../errors';
-import { SecondSignatureAsset, TransactionJSON } from '../transaction_types';
-import { getId, validator } from '../utils';
-import { BaseTransaction, StateStore, StateStorePrepare } from './base';
+import {
+	BaseTransaction,
+	StateStore,
+	StateStorePrepare,
+} from './base_transaction';
+import { SIGNATURE_FEE } from './constants';
+import { TransactionError, TransactionMultiError } from './errors';
+import { SecondSignatureAsset, TransactionJSON } from './transaction_types';
+import { getId, validator } from './utils';
 
 const TRANSACTION_SIGNATURE_TYPE = 1;
