+++ packages/lisk-transactions/src/index.ts
@@ -13,26 +13,25 @@
  *
  */
 import { transfer } from './0_transfer';
+import { TransferTransaction } from './0_transfer_transaction';
 import { registerSecondPassphrase } from './1_register_second_passphrase';
+import { SecondSignatureTransaction } from './1_second_signature_transaction';
+import { DelegateTransaction } from './2_delegate_transaction';
 import { registerDelegate } from './2_register_delegate';
 import { castVotes } from './3_cast_votes';
+import { VoteTransaction } from './3_vote_transaction';
+import { MultisignatureTransaction } from './4_multisignature_transaction';
 import { registerMultisignature } from './4_register_multisignature_account';
 import { createDapp } from './5_create_dapp';
+import { DappTransaction } from './5_dapp_transaction';
+import { InTransferTransaction } from './6_in_transfer_transaction';
+import { OutTransferTransaction } from './7_out_transfer_transaction';
+import { BaseTransaction } from './base_transaction';
 import * as constants from './constants';
 import { createSignatureObject } from './create_signature_object';
 import { TransactionError, TransactionMultiError } from './errors';
 import { Status, TransactionResponse } from './response';
 import { TransactionJSON } from './transaction_types';
-import {
-	BaseTransaction,
-	DelegateTransaction,
-	InTransferTransaction,
-	MultisignatureTransaction,
-	OutTransferTransaction,
-	SecondSignatureTransaction,
-	TransferTransaction,
-	VoteTransaction,
-} from './transactions';
 import * as utils from './utils';
 
 export {
@@ -41,6 +40,7 @@ export {
 	createDapp,
 	MultisignatureTransaction,
 	TransferTransaction,
+	DappTransaction,
 	SecondSignatureTransaction,
 	DelegateTransaction,
 	VoteTransaction,
