--- packages/lisk-transactions/src/transactions/index.ts
@@ -1,23 +1,0 @@
-/*
- * Copyright © 2018 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- *
- */
-export * from './base';
-export * from './0_transfer';
-export * from './1_register_second_passphrase';
-export * from './2_register_delegate';
-export * from './3_vote_transaction';
-export * from './4_register_multisignature';
-export * from './5_dapp_transaction';
-export * from './6_in_transfer_transaction';
-export * from './7_out_transfer_transaction';
