+++ packages/lisk-transactions/test/0_transfer_transaction.ts
@@ -14,16 +14,13 @@
  */
 import * as BigNum from 'browserify-bignum';
 import { expect } from 'chai';
-import { MAX_TRANSACTION_AMOUNT, TRANSFER_FEE } from '../../src/constants';
-import { TransferTransaction } from '../../src/transactions';
-import { Account } from '../../src/transaction_types';
-import { Status } from '../../src/response';
-import { TransactionError } from '../../src/errors';
-import { addTransactionFields, MockStateStore as store } from '../helpers';
-import {
-	validTransferAccount,
-	validTransferTransactions,
-} from '../../fixtures';
+import { MAX_TRANSACTION_AMOUNT, TRANSFER_FEE } from '../src/constants';
+import { TransferTransaction } from '../src/0_transfer_transaction';
+import { Account } from '../src/transaction_types';
+import { Status } from '../src/response';
+import { TransactionError } from '../src/errors';
+import { addTransactionFields, MockStateStore as store } from './helpers';
+import { validTransferAccount, validTransferTransactions } from '../fixtures';
 
 describe('Transfer transaction class', () => {
 	const validTransferTransaction = addTransactionFields(
