+++ packages/lisk-transactions/test/2_delegate_transaction.ts
@@ -13,14 +13,14 @@
  *
  */
 import { expect } from 'chai';
-import { MockStateStore as store } from '../helpers';
-import { DelegateTransaction } from '../../src/transactions';
+import { MockStateStore as store } from './helpers';
+import { DelegateTransaction } from '../src/2_delegate_transaction';
 import {
 	validDelegateAccount,
 	validDelegateTransaction,
 	validTransaction,
-} from '../../fixtures';
-import { Account, TransactionJSON } from '../../src/transaction_types';
+} from '../fixtures';
+import { Account, TransactionJSON } from '../src/transaction_types';
 
 describe('Delegate registration transaction class', () => {
 	let validTestTransaction: DelegateTransaction;
