+++ packages/lisk-transactions/test/3_vote_transaction.ts
@@ -13,12 +13,12 @@
  *
  */
 import { expect } from 'chai';
-import { MockStateStore as store } from '../helpers';
-import { VoteTransaction } from '../../src/transactions';
-import { validVoteTransactions } from '../../fixtures';
-import { TransactionJSON } from '../../src/transaction_types';
-import { Status } from '../../src/response';
-import { generateRandomPublicKeys } from '../helpers/cryptography';
+import { MockStateStore as store } from './helpers';
+import { VoteTransaction } from '../src/3_vote_transaction';
+import { validVoteTransactions } from '../fixtures';
+import { TransactionJSON } from '../src/transaction_types';
+import { Status } from '../src/response';
+import { generateRandomPublicKeys } from './helpers/cryptography';
 
 describe('Vote transaction class', () => {
 	let validTestTransaction: VoteTransaction;
