+++ packages/lisk-transactions/test/utils/verify.ts
@@ -23,7 +23,7 @@ import {
 } from '../../src/utils';
 import * as validator from '../../src/utils/sign_and_verify';
 import { TransactionError, TransactionPendingError } from '../../src/errors';
-import { MultisignatureStatus } from '../../src/transactions';
+import { MultisignatureStatus } from '../../src/base_transaction';
 
 describe('#verify', () => {
 	const defaultId = '4838520211125422557';
